requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
cffi==1.14.5
colorama==0.4.4
coverage==5.5
cryptography==3.4.6
Django==3.1.7
django-filter==2.4.0
django-rest-knox==4.1.0
django-rest-registration==0.6.0
djangorestframework==3.12.2
Faker==6.5.0
iniconfig==1.1.1
mixer==7.1.2
model-bakery==1.2.1
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
